.gitignore: permit advent.ini file, but root directory.
We have splited into seperate directories for which serves
each year's calendars. So we do not have a global config file,
which could located in the root directory of the branch if not.
We still have config files in each sub-directories. So, ini file
names in gitignore list should be prefixed with /.